Switch mouse picking to use UINT color format
Changed color format from VK_FORMAT_R8G8B8A8_UNORM to VK_FORMAT_R8G8B8A8_UINT in pipeline, renderer, and image resource logic. Updated MousePicking.frag shader to output uvec4 and removed normalization. Adjusted MousePicker.cpp to correctly extract objectID from uint8_t pixel data. This improves precision and consistency for object picking.

+++ source/components/renderingManager/renderer3D/MousePicker.cpp
@@ -136,7 +136,9 @@ namespace vke {
     m_logicalDevice->doMappedMemoryOperation(stagingBufferMemory, [&objectID](void* data) {
       const uint8_t* pixel = static_cast<uint8_t*>(data);
 
-      objectID = pixel[0] << 16 | pixel[1] << 8 | pixel[2];
+      objectID = static_cast<uint32_t>(pixel[0]) << 16 |
+                 static_cast<uint32_t>(pixel[1]) << 8 |
+                 static_cast<uint32_t>(pixel[2]);
     });
 
     return objectID;
